Fix cache cleaning bug

+++ repositories/object_cache_repository.go
@@ -7,10 +7,15 @@ import (
 	"path/filepath"
 )
 
+var (
+	ErrRemoveFileFromCache = fmt.Errorf("error removing file from cache")
+)
+
 type ObjectCacheRepository struct {
-	resolver  func(id string, writer io.Writer) (err error)
-	readPath  string
-	writePath string
+	resolver       func(id string, writer io.Writer) (err error)
+	readPath       string
+	writePath      string
+	committingList map[string]struct{}
 }
 
 func NewObjectCacheRepository(path string) ObjectCacheRepository {
@@ -20,8 +25,9 @@ func NewObjectCacheRepository(path string) ObjectCacheRepository {
 		panic(err)
 	}
 	return ObjectCacheRepository{
-		readPath:  path,
-		writePath: writePath,
+		readPath:       path,
+		writePath:      writePath,
+		committingList: make(map[string]struct{}),
 	}
 }
 
@@ -118,8 +124,12 @@ func (o *ObjectCacheRepository) IsInCache(id string) (is bool) {
 }
 
 func (o *ObjectCacheRepository) Flush(id string) (err error) {
+	delete(o.committingList, id)
 	p := filepath.Join(o.writePath, id)
 	err = os.Remove(p)
+	if err != nil {
+		return
+	}
 	return
 }
 
@@ -147,12 +157,21 @@ func (o *ObjectCacheRepository) resolverFile(id string) (err error) {
 // It returns an error if the removal operation fails.
 // If the object is not in the cache, it returns nil (no error).
 func (o *ObjectCacheRepository) RemoveFromCache(id string) error {
+	// If the object is in the list of committed objects, we should wait for it to be committed.
+	_, committed := o.committingList[id]
+	if committed {
+		return nil
+	}
+
 	p := filepath.Join(o.readPath, id)
 	if _, err := os.Stat(p); os.IsNotExist(err) {
 		return nil
 	}
 	err := os.Remove(p)
-	return err
+	if err != nil {
+		return ErrRemoveFileFromCache
+	}
+	return nil
 }
 
 // IsOpenForWrite returns true if the object is in the write cache.
@@ -161,5 +180,11 @@ func (o *ObjectCacheRepository) IsOpenForWrite(id string) bool {
 	if _, err := os.Stat(p); os.IsNotExist(err) {
 		return false
 	}
-	return true
+	_, committed := o.committingList[id]
+	return !committed
+}
+
+// AdToCommitting marks the object as committed in the write cache.
+func (o *ObjectCacheRepository) AdToCommitting(id string) {
+	o.committingList[id] = struct{}{}
 }
